+++ src/screen/screenType.py
@@ -6,4 +6,5 @@ class ScreenType:
     STATS_SCREEN = "stats_screen"
     INVENTORY_SCREEN = "inventory_screen"
     CONFIG_SCREEN = "config_screen"
+    SAVE_SELECTION_SCREEN = "save_selection_screen"
     NONE = "none_screen"